Rename answer mutation to provideAnswereToPollQuestion
Also: Provide better error handling
